emotion-detection-service/app/requirements.txt
numpy==1.24.3
onnxruntime==1.15.0
onnx==1.14.0
opencv-python==4.7.0.72
mediapipe==0.10.1
paho-mqtt==1.6.1